Use modern cmake

+++ src/coin.h
@@ -1,15 +1,16 @@
-#include <stdlib.h>
-#include <string.h>
-#include "util.h"
 #ifndef COIN_H
 #define COIN_H
+#include <stdlib.h>
+#include <string.h>
+
+#include "util.h"
 
 typedef struct {
     size_t start;
     size_t limit;
     char* convert;
     char** symbols;
-    int specific;
+    size_t specific;
     int global;
     char* portfolio;
     int color_enabled;
